Fix entire buffer rendering bold while typing
For whatever reason, TSError is partially applied to the entire buffer
contents when you're typing, and when there is a syntax error (such as
when you haven't closed a parenthesis yet or something) the entire
document keeps flashing bold as you type.

It was driving me mad, but I finally went and tracked it down to TSError
-- removing the bold from this solves the problem. Interestingly the
pink text colour was only applied to the trailing token and not the
entire document like the bold was, so probably is a bug somewhere
upstream.