Reduce drift and duplication between TS and Base
Treesitter redefines a lot of stuff that is already defined with the
built-in vim equivalents. We should refactor the theme definitions so we
can reference the base values directly to assign the exact same thing to
the equivalent treesitter groups.

This improves consistency between TS and non-TS syntaxes, and lets us do
some other stuff to reduce duplication and potential drift, like add
"presets" for commonly-used style combinations.