Make cyan, green, and yellow a bit bolder
The yellow/blue/red colours started to feel a bit too "kid's toy". This
makes the yellow a bit more orange to make it bolder and less primary;
cyan takes a similar treatment, pushing it a tiny bit closer to blue;
and finally green takes the biggest change, bringing it a lot darker and
a bit cooler to make it blend in more.

Really liking how this is bringing the theme together.