Make line numbers separate from file content
This separates the LineNr more from the body of the file by giving it
a very light bg and lightening the text colour a bit. This not only
helps visualise where the beginning of a line is (the empty space can
make things look indented when they're not) but also separates the line
numbers so they're less distracting when you're just trying to read the
file.