Merge red with orange
These are extremely close to one another and not really worth the
difference. Let's just axe red as orange fits in better anyway.

The only weird one was DiffChange was for some reason orange, so we move
that to be yellow which makes more sense anyway, and then the red for
DiffRemove can become orange.