Cooler, simpler
This removes the majority of the orange and removes colour from some of
the most common bits, leaving highlighting only for functions, special
tokens, and control flow.

To me this is more purely what syntax highlighting is really for:
highlighting critical parts of the code, but without making it look like
a circus. This theme was pretty subdued before, but these recent changes
have struck a chord with me.