Lunarised 2.0: Eclipse
This redoes the entire palette to be more cohesive. Colours were
inspired by astrophotography of a lunar eclipse and rich reds, oranges,
and pale blues and purples that come from that.

This abandons the original Solarized palette that the previous colours
were based off of in favour of all-new custom colours.

This is not only more muted, but feels more "designed" -- the previous
colours always felt a little too bright and clashed a little too much in
certain places. The more muted moon-tones not only fit the name better,
but also feel much more natural together.